Change --order argument parsing to use the fields from FIELD_MAP as named in "pypinfo --help" (#9)
* DEV: Harmonize --order field names

Previously, the field names passed to `order` had to be those
of the actual BigQuery database, and not those displayed
with `pypinfo --help`. Now the appropriate field substitution
is made when the `order` argument is passed into `build_query`.

* DEV: Change to allow any ordering field

Now any value can be passed to `-o` and, if it's not a key in
FIELD_MAP it's just passed in verbatim. This still works when
no `-o` is passed, since the default `None` is also passed
through verbatim.